pyproject.toml
[project]
name = "video-subtitle"
version = "0.1.0"
requires-python = ">=3.11,<4.0"
dependencies = [
    "faster-whisper>=1.0.0",
    "ffmpeg-python>=0.2.0",
    "click>=8.1.7",
    "pytest>=7.4.0"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
